+++ test/test-autodownload.sh
@@ -5,9 +5,8 @@ set -e
 TITLE="MUROTAL DOWNLOADER SCRIPT"    # script name
 VER="2.2"                            # script version
 
-PATH_MUROTAL="/Users/dwidenni/Documents/murotal/1000_Misyari_Rasyid"
-DOWNLOAD_PATH="https://ia801204.us.archive.org/17/items/Misyari_Rasyid_per_Juz"
-ID_CODE="10"
+ID_CODE=$1
+PATH_MUROTAL=$2
 
 COL_RED="\033[22;31m"
 COL_GREEN="\033[22;32m"
@@ -78,6 +77,60 @@ download_murotal_perjuz() {
   done
 }
 
+selected_download() {
+  case $ID_CODE in
+    10 | 1000)
+      export PATH_MUROTAL="$PATH_MUROTAL/1000_Misyari_Rasyid"
+      export DOWNLOAD_PATH="https://ia801204.us.archive.org/17/items/Misyari_Rasyid_per_Juz"
+      ;;
+
+    12 | 1200)
+      export PATH_MUROTAL="$PATH_MUROTAL/1200_Saad_AlGhamidi"
+      export DOWNLOAD_PATH="https://ia903200.us.archive.org/20/items/SaadAl-GhamidiPerJuz"
+      ;;
+
+    14 | 1400)
+      export PATH_MUROTAL="$PATH_MUROTAL/1400_Abdurrahman_AsSudais"
+      export DOWNLOAD_PATH="https://ia801904.us.archive.org/14/items/AbdurrahmanSudaisPerJuz"
+      ;;
+
+    16 | 1600)
+      export PATH_MUROTAL="$PATH_MUROTAL/1600_Hany_ArRifai"
+      export DOWNLOAD_PATH="https://ia801207.us.archive.org/10/items/HanyAr-RifaiPerJuz"
+      ;;
+
+    18 | 1800)
+      export PATH_MUROTAL="$PATH_MUROTAL/1800_Yasser_AlDosari"
+      export DOWNLOAD_PATH="https://ia803203.us.archive.org/26/items/YasserAlDosariPerJuz"
+      ;;
+
+    20 | 2000)
+      export PATH_MUROTAL="$PATH_MUROTAL/2000_Saud_AlShuraim"
+      export DOWNLOAD_PATH="https://ia903208.us.archive.org/20/items/SaudAlShuraimPerJuz"
+      ;;
+
+    22 | 2200)
+      export PATH_MUROTAL="$PATH_MUROTAL/2200_Nasser_AlQatami"
+      export DOWNLOAD_PATH="https://ia800400.us.archive.org/23/items/Nasseralqatamieveryjuz"
+      ;;
+
+    24 | 2400)
+      export PATH_MUROTAL="$PATH_MUROTAL/2400_Abdul_Basit_AbdulSamad"
+      export DOWNLOAD_PATH="https://ia903200.us.archive.org/23/items/AbdulBasitAbdulsamadPerJuz"
+      ;;
+
+    *)
+      echo ""
+      echo ">> Unknown MID (Murotal ID)"
+      echo ">> Set to default MID: 10"
+      export PATH_MUROTAL="$PATH_MUROTAL/1000_Misyari_Rasyid"
+      export DOWNLOAD_PATH="https://ia801204.us.archive.org/17/items/Misyari_Rasyid_per_Juz"
+      echo ""
+      ;;
+  esac
+  download_murotal_perjuz
+}
+
 download_below10() {
   JUZ="0$1"
   DOWNLOAD_LINK="$DOWNLOAD_PATH/$JUZ.mp3"
@@ -119,9 +172,9 @@ skip_exists() {
 
 main() {
   header
-  download_murotal_perjuz
+  selected_download $1 $2
   footer
 }
 
 ### START HERE ###
-main $@
+main $1 $2